# renderer: rewrite glyph cache