# Rewriting the glyph cache

This has been planned by @Guldoman and me for a while now since the current GlyphSet-based cache cannot properly support more advanced glyph manipulation (ligatures anyone?) and is very wasteful in terms of memory as it allocates a relatively huge struct, approximately 0x100000 times. This rewrite splits the cache into two:

### Mapping codepoints to glyph IDs: `CharMap`

This is a 2D array that maps each codepoint to glyph IDs. Glyph IDs are assumed to be uint16 (according to OpenType, I've not known other formats that have a larger glyph ID). The rows of this map are dynamically allocated for efficiency, and when fully allocated, it's roughly 4,457,468 bytes. This map can be replaced with something else for more features, like shaping.

### Mapping glyph IDs to glyphs: `GlyphMap`

This is the main part of the rewrite. A GlyphMap contains a map of GlyphMetrics similar to how CharMap works. This map adds up to roughly 2,098,176 bytes. A GlyphMap also contains a list of GlyphAtlas per (subpixel) bitmap index (1 if grayscale AA, 3 if subpixel AA), mapped by their width.

A GlyphAtlas represents a list of SDL_Surfaces with the same width. Each glyph is stacked on the surface vertically. The GlyphAtlas acts as a bump allocator; when someone tries to store something bigger than the current surface can handle, it allocates a new surface with the appropriate size and stores it there. If the current surface has sufficient space, it stores the bitmap there directly.
The GlyphSet will try to find a GlyphAtlas with a suitable width or allocate a new one.

The renderer no longer renders an entire GlyphSet (256 characters) at once for one glyph. Glyphs are packed tightly together whenever possible. Calls to FT_Get_Char_Index are also cached, which would improve performance a bit.

### Other changes

Aside from the GlyphSet cache, the tab character handling has been reworked _a bit_ not to directly set the values in the GlyphMetric. The FT_Stream is now properly freed. The check_alloc function is modified to display file and line number (which can be useful), and an internal function called `ren_font_dump()` is added to dump a font's contents into a series of BMP files for inspection. This is guarded behind a preprocessor macro called `RENDERER_DEBUG` and is disabled by default. This function is intended to be used inside a debugger.

### Benchmarks

This is the most exciting part. Comparing this PR to master:

| My PR | Master |
| ----- | ------ |
| <pre>*** get-width-fira-sans :    5.267ms  31.60%<br>*** get-width-fira-sans-grayscale :    2.539ms  15.23%<br>*** get-width-cjk    :   17.946ms 107.68%<br>*** get-width-cjk-grayscale :   11.705ms  70.23%</pre> | <pre>*** get-width-fira-sans :   47.054ms 282.32%<br>*** get-width-fira-sans-grayscale :   15.333ms  92.00%<br>*** get-width-cjk    : 6337.212ms 38023.27%<br>*** get-width-cjk-grayscale : 1207.724ms 7246.34%</pre> |

Compared to my previous PR #1543:

| My PR | #1543 |
| ----- | ----- |
| <pre>*** get-width-fira-sans :    5.267ms  31.60%<br>*** get-width-fira-sans-grayscale :    2.539ms  15.23%<br>*** get-width-cjk    :   17.946ms 107.68%<br>*** get-width-cjk-grayscale :   11.705ms  70.23%</pre> | <pre>*** get-width-fira-sans :    6.434ms  38.60%<br>*** get-width-fira-sans-grayscale :    5.113ms  30.68%<br>*** get-width-cjk    :  502.446ms 3014.67%<br>*** get-width-cjk-grayscale :  419.128ms 2514.77%</pre> |

As you can see, the time is inconsistent across even the same run, but this PR demonstrates **a clear increase** compared to master and #1543.

<details>
<summary>Code</summary>

```lua
local common = require "core.common"

common.bench("get-width-fira-sans", function()
  local font = renderer.font.load(DATADIR .. "/fonts/FiraSans-Regular.ttf", 14 * SCALE)
  for i = 1, 100 do
    font:get_width(utf8.char(32 + i))
  end
end)

common.bench("get-width-fira-sans-grayscale", function()
  local font = renderer.font.load(DATADIR .. "/fonts/FiraSans-Regular.ttf", 14 * SCALE, { antialiasing = "grayscale" })
  for i = 1, 100 do
    font:get_width(utf8.char(32 + i))
  end
end)

common.bench("get-width-cjk", function()
  local font = renderer.font.load(DATADIR .. "/fonts/NotoSansCJKsc-Regular.otf", 14 * SCALE)
  for i = 1, 100 do
    font:get_width(utf8.char(1000 + i + i * 256))
  end
end)

common.bench("get-width-cjk-grayscale", function()
  local font = renderer.font.load(DATADIR .. "/fonts/NotoSansCJKsc-Regular.otf", 14 * SCALE, { antialiasing = "grayscale" })
  for i = 1, 100 do
    font:get_width(utf8.char(1000 + i + i * 256))
  end
end)
```
</details>

### Real-life tests

This PR:

https://github.com/lite-xl/lite-xl/assets/20792268/9fd1ed38-4ba8-4c0f-8f47-b39780ed8115

#1543:

https://github.com/lite-xl/lite-xl/assets/20792268/52144e6f-1b2d-431f-b268-75459c51d02f

Master:

https://github.com/lite-xl/lite-xl/assets/20792268/3ee957a9-e6bc-4492-9efd-d2067ea33ff8